+++ test/test-bank.cpp
@@ -7,34 +7,34 @@ using namespace qc::primitives;
 
 TEST(Bank, general)
 {
-    qc::Bank<int> bank{};
+    qc::Bank<u32> bank{};
     ASSERT_EQ(0u, bank.capacity());
     ASSERT_EQ(0u, bank.size());
     ASSERT_TRUE(bank.empty());
 
-    ASSERT_EQ(0u, bank.create(0).i);
-    ASSERT_EQ(0, bank[0u]);
+    ASSERT_EQ(0u, bank.create(0u).i);
+    ASSERT_EQ(0u, bank[0u]);
     ASSERT_EQ(16u, bank.capacity());
     ASSERT_EQ(1u, bank.size());
     ASSERT_FALSE(bank.empty());
 
-    for (int i{1}; i < 16; ++i)
+    for (u32 i{1u}; i < 16u; ++i)
     {
         ASSERT_EQ(i, bank.create(i).i);
-        ASSERT_EQ(i, bank[uint(i)]);
+        ASSERT_EQ(i, bank[i]);
         ASSERT_EQ(16u, bank.capacity());
-        ASSERT_EQ(i + 1, bank.size());
+        ASSERT_EQ(i + 1u, bank.size());
     }
 
-    ASSERT_EQ(16u, bank.create(16).i);
+    ASSERT_EQ(16u, bank.create(16u).i);
     ASSERT_EQ(32u, bank.capacity());
     ASSERT_EQ(17u, bank.size());
 
-    for (int i{0}; i < 16; ++i)
+    for (u32 i{0u}; i < 16u; ++i)
     {
-        bank.destroy(uint(i));
+        bank.destroy(i);
         ASSERT_EQ(32u, bank.capacity());
-        ASSERT_EQ(16 - i, bank.size());
+        ASSERT_EQ(16u - i, bank.size());
     }
 
     ASSERT_EQ(32u, bank.capacity());
@@ -50,40 +50,40 @@ TEST(Bank, general)
 
 TEST(Bank, generalNontrivial)
 {
-    qc::Bank<std::unique_ptr<int>> bank{};
+    qc::Bank<std::unique_ptr<u32>> bank{};
     ASSERT_EQ(0u, bank.capacity());
     ASSERT_EQ(0u, bank.size());
     ASSERT_TRUE(bank.empty());
 
-    ASSERT_EQ(0u, bank.create(new int{0}).i);
-    ASSERT_EQ(0, *bank[0u]);
+    ASSERT_EQ(0u, bank.create(new u32{0u}).i);
+    ASSERT_EQ(0u, *bank[0u]);
     ASSERT_EQ(16u, bank.capacity());
     ASSERT_EQ(1u, bank.size());
     ASSERT_FALSE(bank.empty());
 
-    for (int i{1}; i < 16; ++i)
+    for (u32 i{1u}; i < 16u; ++i)
     {
-        ASSERT_EQ(i, bank.create(new int{i}).i);
-        ASSERT_EQ(i, *bank[uint(i)]);
+        ASSERT_EQ(i, bank.create(new u32{i}).i);
+        ASSERT_EQ(i, *bank[i]);
         ASSERT_EQ(16u, bank.capacity());
-        ASSERT_EQ(i + 1, bank.size());
+        ASSERT_EQ(i + 1u, bank.size());
     }
 
-    ASSERT_EQ(16, bank.create(new int{16}).i);
+    ASSERT_EQ(16u, bank.create(new u32{16u}).i);
     ASSERT_EQ(32u, bank.capacity());
     ASSERT_EQ(17u, bank.size());
 
-    for (int i{0}; i < 16; ++i)
+    for (u32 i{0u}; i < 16u; ++i)
     {
-        bank.destroy(uint(i));
+        bank.destroy(i);
         ASSERT_EQ(32u, bank.capacity());
-        ASSERT_EQ(16 - i, bank.size());
+        ASSERT_EQ(16u - i, bank.size());
     }
 
     ASSERT_EQ(32u, bank.capacity());
     ASSERT_EQ(1u, bank.size());
     ASSERT_FALSE(bank.empty());
-    ASSERT_EQ(16, *bank[16u]);
+    ASSERT_EQ(16u, *bank[16u]);
 
     bank.destroy(16u);
     ASSERT_EQ(32u, bank.capacity());
@@ -93,7 +93,7 @@ TEST(Bank, generalNontrivial)
 
 TEST(Bank, reserve)
 {
-    qc::Bank<int> bank{};
+    qc::Bank<u32> bank{};
     ASSERT_EQ(0u, bank.capacity());
     ASSERT_EQ(0u, bank.size());
 
@@ -120,26 +120,26 @@ TEST(Bank, reserve)
 
 TEST(Bank, freeOrder)
 {
-    qc::Bank<int> bank{};
+    qc::Bank<u32> bank{};
 
-    ASSERT_EQ(0u, bank.create(0).i);
-    ASSERT_EQ(0, bank[0u]);
+    ASSERT_EQ(0u, bank.create(0u).i);
+    ASSERT_EQ(0u, bank[0u]);
 
-    ASSERT_EQ(1u, bank.create(1).i);
-    ASSERT_EQ(1, bank[1u]);
+    ASSERT_EQ(1u, bank.create(1u).i);
+    ASSERT_EQ(1u, bank[1u]);
 
     bank.destroy(0u);
 
-    ASSERT_EQ(0u, bank.create(2).i);
-    ASSERT_EQ(2, bank[0u]);
+    ASSERT_EQ(0u, bank.create(2u).i);
+    ASSERT_EQ(2u, bank[0u]);
 
-    ASSERT_EQ(2u, bank.create(3).i);
-    ASSERT_EQ(3, bank[2u]);
+    ASSERT_EQ(2u, bank.create(3u).i);
+    ASSERT_EQ(3u, bank[2u]);
 
     bank.destroy(2u);
 
-    ASSERT_EQ(2u, bank.create(4).i);
-    ASSERT_EQ(4, bank[2u]);
+    ASSERT_EQ(2u, bank.create(4u).i);
+    ASSERT_EQ(4u, bank[2u]);
 
     ASSERT_EQ(3u, bank.size());
 
@@ -147,33 +147,33 @@ TEST(Bank, freeOrder)
     bank.reserve(32u);
     ASSERT_EQ(32u, bank.capacity());
 
-    for (int i{0}; i < 16; ++i)
+    for (u32 i{0u}; i < 16u; ++i)
     {
-        ASSERT_EQ(16u + u32(i), bank.create(5 + i).i);
-        ASSERT_EQ(5 + i, bank[16u + u32(i)]);
+        ASSERT_EQ(16u + i, bank.create(5u + i).i);
+        ASSERT_EQ(5u + i, bank[16u + i]);
     }
 
     ASSERT_EQ(19u, bank.size());
 
-    for (int i{0}; i < 13; ++i)
+    for (u32 i{0u}; i < 13u; ++i)
     {
-        ASSERT_EQ(3u + u32(i), bank.create(21 + i).i);
-        ASSERT_EQ(21 + i, bank[3u + u32(i)]);
+        ASSERT_EQ(3u + i, bank.create(21u + i).i);
+        ASSERT_EQ(21u + i, bank[3u + i]);
     }
 
     ASSERT_EQ(32u, bank.size());
 
-    for (int i{0}; i < 32; ++i)
+    for (u32 i{0u}; i < 32u; ++i)
     {
-        bank.destroy(uint(31 - i));
+        bank.destroy(31u - i);
     }
 
     ASSERT_EQ(0u, bank.size());
 
-    for (int i{0}; i < 32; ++i)
+    for (u32 i{0u}; i < 32u; ++i)
     {
-        ASSERT_EQ(u32(i), bank.create(100 + i).i);
-        ASSERT_EQ(100 + i, bank[uint(i)]);
+        ASSERT_EQ(i, bank.create(100u + i).i);
+        ASSERT_EQ(100u + i, bank[i]);
     }
 
     ASSERT_EQ(32u, bank.size());
@@ -182,14 +182,14 @@ TEST(Bank, freeOrder)
 
 TEST(Bank, clear)
 {
-    qc::Bank<int> bank{};
+    qc::Bank<u32> bank{};
 
     bank.clear();
     ASSERT_TRUE(bank.empty());
     ASSERT_EQ(0u, bank.capacity());
 
-    auto [e0, i0]{bank.create(0)};
-    ASSERT_EQ(0, e0);
+    auto [e0, i0]{bank.create(0u)};
+    ASSERT_EQ(0u, e0);
     ASSERT_EQ(0u, i0);
     ASSERT_EQ(1u, bank.size());
     ASSERT_EQ(16u, bank.capacity());
@@ -198,14 +198,14 @@ TEST(Bank, clear)
     ASSERT_TRUE(bank.empty());
     ASSERT_EQ(16u, bank.capacity());
 
-    auto [e1, i1]{bank.create(1)};
-    ASSERT_EQ(1, e1);
+    auto [e1, i1]{bank.create(1u)};
+    ASSERT_EQ(1u, e1);
     ASSERT_EQ(0u, i1);
-    auto [e2, i2]{bank.create(2)};
-    ASSERT_EQ(2, e2);
+    auto [e2, i2]{bank.create(2u)};
+    ASSERT_EQ(2u, e2);
     ASSERT_EQ(1u, i2);
-    auto [e3, i3]{bank.create(3)};
-    ASSERT_EQ(3, e3);
+    auto [e3, i3]{bank.create(3u)};
+    ASSERT_EQ(3u, e3);
     ASSERT_EQ(2u, i3);
     ASSERT_EQ(16u, bank.capacity());
 
@@ -213,14 +213,14 @@ TEST(Bank, clear)
     ASSERT_TRUE(bank.empty());
     ASSERT_EQ(16u, bank.capacity());
 
-    auto [e4, i4]{bank.create(4)};
-    ASSERT_EQ(4, e4);
+    auto [e4, i4]{bank.create(4u)};
+    ASSERT_EQ(4u, e4);
     ASSERT_EQ(0u, i4);
-    auto [e5, i5]{bank.create(5)};
-    ASSERT_EQ(5, e5);
+    auto [e5, i5]{bank.create(5u)};
+    ASSERT_EQ(5u, e5);
     ASSERT_EQ(1u, i5);
-    auto [e6, i6]{bank.create(6)};
-    ASSERT_EQ(6, e6);
+    auto [e6, i6]{bank.create(6u)};
+    ASSERT_EQ(6u, e6);
     ASSERT_EQ(2u, i6);
     ASSERT_EQ(16u, bank.capacity());
 
@@ -231,14 +231,14 @@ TEST(Bank, clear)
 
 TEST(Bank, clearNontrivial)
 {
-    qc::Bank<std::unique_ptr<int>> bank{};
+    qc::Bank<std::unique_ptr<u32>> bank{};
 
     bank.clear();
     ASSERT_TRUE(bank.empty());
     ASSERT_EQ(0u, bank.capacity());
 
-    auto [e0, i0]{bank.create(new int{0})};
-    ASSERT_EQ(0, *e0);
+    auto [e0, i0]{bank.create(new u32{0u})};
+    ASSERT_EQ(0u, *e0);
     ASSERT_EQ(0u, i0);
     ASSERT_EQ(1u, bank.size());
     ASSERT_EQ(16u, bank.capacity());
@@ -247,14 +247,14 @@ TEST(Bank, clearNontrivial)
     ASSERT_TRUE(bank.empty());
     ASSERT_EQ(16u, bank.capacity());
 
-    auto [e1, i1]{bank.create(new int{1})};
-    ASSERT_EQ(1, *e1);
+    auto [e1, i1]{bank.create(new u32{1u})};
+    ASSERT_EQ(1u, *e1);
     ASSERT_EQ(0u, i1);
-    auto [e2, i2]{bank.create(new int{2})};
-    ASSERT_EQ(2, *e2);
+    auto [e2, i2]{bank.create(new u32{2u})};
+    ASSERT_EQ(2u, *e2);
     ASSERT_EQ(1u, i2);
-    auto [e3, i3]{bank.create(new int{3})};
-    ASSERT_EQ(3, *e3);
+    auto [e3, i3]{bank.create(new u32{3u})};
+    ASSERT_EQ(3u, *e3);
     ASSERT_EQ(2u, i3);
     ASSERT_EQ(16u, bank.capacity());
